MelanomaRecognition/telegram_bot/requirements.txt
torch
torchvision
Pillow
python-telegram-bot
python-dotenv
numpy

